Add files via upload
This is a program that you can use through python on your computer for games. if you are playing a FPS game then use this and it basically clicks really fast for you when you click the right click button. For example, if you play shellshock.io, go to pistol, turn this on, and right click and it will empty your whole clip in about 2 seconds which is insanely quick.  Anyways hope you enjoy - Fiverr :